refactor(sync): extract sync_engine/block_source.rs
Move MemoryBlockSource, MemoryBlockSourceError, and the BlockSource
trait impl out of sync_engine/mod.rs into a dedicated block_source
submodule. The private `blocks` field is now reached through a
`MemoryBlockSource::new(blocks)` constructor so the struct field can
stay out of the super-module's surface.

Visibility is `pub(super)` — only sync_engine/mod.rs constructs one
(inside `download_blocks`) and passes it straight to
`scan_cached_blocks`.

mod.rs no longer needs `std::fmt`, `chain::BlockSource`, or the
`proto::compact_formats::CompactBlock` import directly, so those are
dropped from the top-of-file import block.

23 tests pass.
